chatbot/requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain
openai
langchain-community
langchain-core
qdrant-client